+++ android/library/src/main/java/com/liferay/mobile/screens/auth/login/LoginScreenlet.java
@@ -116,12 +116,12 @@ public class LoginScreenlet extends BaseScreenlet<LoginViewModel, BaseLoginInter
 		}
 	}
 
-	public void setListener(LoginListener listener) {
-		this.listener = listener;
-	}
-
 	public LoginListener getListener() {
 		return listener;
+	}
+
+	public void setListener(LoginListener listener) {
+		this.listener = listener;
 	}
 
 	public BasicAuthMethod getAuthMethod() {
@@ -216,8 +216,7 @@ public class LoginScreenlet extends BaseScreenlet<LoginViewModel, BaseLoginInter
 	protected BaseLoginInteractor createInteractor(String actionName) {
 		if (authenticationType.equals(AuthenticationType.COOKIE)) {
 			return new LoginCookieInteractor();
-		}
-		else if (authenticationType.equals(AuthenticationType.OAUTH)) {
+		} else if (authenticationType.equals(AuthenticationType.OAUTH)) {
 			LoginOAuthInteractor oauthInteractor = new LoginOAuthInteractor();
 
 			OAuthConfig config =
@@ -226,8 +225,7 @@ public class LoginScreenlet extends BaseScreenlet<LoginViewModel, BaseLoginInter
 			oauthInteractor.setOAuthConfig(config);
 
 			return oauthInteractor;
-		}
-		else {
+		} else {
 			return new LoginBasicInteractor();
 		}
 	}
@@ -240,7 +238,7 @@ public class LoginScreenlet extends BaseScreenlet<LoginViewModel, BaseLoginInter
 		} else if (AuthenticationType.BASIC.equals(authenticationType)) {
 			LoginViewModel viewModel = getViewModel();
 			interactor.start(viewModel.getLogin(), viewModel.getPassword(), viewModel.getBasicAuthMethod());
-		} else if(AuthenticationType.OAUTH.equals(authenticationType)) {
+		} else if (AuthenticationType.OAUTH.equals(authenticationType)) {
 			LoginOAuthInteractor oauthInteractor = (LoginOAuthInteractor) interactor;
 			Intent intent = new Intent(getContext(), OAuthActivity.class);
 			intent.putExtra(OAuthActivity.EXTRA_OAUTH_CONFIG, oauthInteractor.getOAuthConfig());