+++ android/samples/westerosemployees/src/main/java/com/liferay/mobile/screens/westerosemployees/Views/BlogsCard.java
@@ -60,7 +60,6 @@ public class BlogsCard extends CommentsRatingsCard implements BaseListListener<A
 
 	@Override
 	public void onListItemSelected(AssetEntry element, View view) {
-
 		blogDisplayScreenlet.load(element);
 
 		initializeRatingsAndComments("com.liferay.blogs.kernel.model.BlogsEntry", element.getClassPK());